Fix Bug

+++ util/createFolders.py
@@ -8,9 +8,10 @@ if not os.path.exists(os.path.join(directory, 'models')):
 if not os.path.exists(os.path.join(directory, 'output', 'svg')):
     os.makedirs(os.path.join(directory, 'output', 'svg'))
     
-if True:
-    shutil.rmtree(os.path.join(directory, 'cache'))
-    
+if os.path.exists(os.path.join(directory, 'cache')):
+    if True:
+        shutil.rmtree(os.path.join(directory, 'cache'))
+        
 if not os.path.exists(os.path.join(directory, 'cache')):
     os.makedirs(os.path.join(directory, 'cache'))
